Added pydantic files
